Makefile: solve issue with make -jx flag
Change gen_bin phony target with real $(AT91BOOTSTRAP) binary file. That allow
to make the size checking target depend on it. It wait for it before being executed.
That solves the bug seen while compiling with make -jx directive which was
shown as :
xxxxxxxxx.bin No such file or directory
Size of xxxxxxxxx.bin is  bytes
/bin/sh: line 2: [: : integer expression expected
with error 2

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>