libgcc: remove the dependence to the toolchain's libgcc
For some toolchains, there are some compatibility issue with the toolchain's libgcc,

Because after importing this lib, for some toolchains such as the toolchain
comes with ubuntu 12.04(arm-linux-gnueabi-gcc),it builds OK without
warning and error, but for ek with "926ej-s" core, it fails to run on the board.

for some toolchains such as from Sourcery CodeBench Lite 2012.03-57,
it run on sama5d3xek, no print message output from DBGU. and so on.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>
Tested-by: Josh Wu <Josh.wu@atmel.com>