Add the div() and mod() division function.
Add the div() and mod() division function, to eliminate the below
compilation error which arose after removing the dependence to libgcc:

	undefined reference to `__aeabi_uidiv'
	undefined reference to `__aeabi_uidivmod'

For the division in udelay() function, since our own division function
which spends much run time, causes the delay time inaccuracy. So,
using shifting, change to implement the division.
In spite of changing "1000" to "1024" which causes some inaccuacy,
but it is acceptable for our application.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>
Tested-by: Josh Wu <Josh.wu@atmel.com>