sama5d4ek: nand: increase timing of SMC: Twb according to datasheet
current we set the Twb (Write enable signal to ready/busy signal) to
3 cycle, that means in 170MHz is 33ns.

In such setting, we meet a issue: sometimes the NFC sram read data is
not correct. And different rev.D board has different error frequence.

The root cause is the ready/busy is no valid since we set Twb too small.

According to MT29F4G08ABAEAWP datasheet, Twb max value is 100ns in
3.3V.

So the reasonable value should set Twb as 9 cycles. In 170MHz that is
about 100ns. But the Twb cycle = TWB[bit 3] * 64 + WWB[bit 2:0] clocks.
So we only can set it as 7 cycles or 64 cycles. Cannot set as 9 cycles.
This patch set it as 7, means Twb is 77ns in 170MHz.

Signed-off-by: Josh Wu <josh.wu@atmel.com>