Makefile: use defconfig instead of oldconfig when do <board>_defconfig
As oldconfig will make all old symbol as default. For the all new symbols,
it will ask to user input.

For defconfig, it will set all new symbols to default value instead of ask
user for input.

As we want the <board>_defconfig is a minimum one. That means the default value
can be set automatically. So we need the defconfig.

This patch make the 'make defconfig' use .config as default config.
And .config is copied from the <board>_defconfig

Signed-off-by: Josh Wu <josh.wu@atmel.com>