L2CC: split cache operation to allow proper setup
We split cache initialization in two functions:
- one for setting the cache configuration values (mostly prefetch and power
  management)
- one to actually enable it: this one only will be called from the NWd if
  needed.

This allows to configure the L2 Cache even when we only use the SoC in SWd.
This way, later in the boot sequence when the cache is enabled (by Linux for
instance) its configuration values are already set correctly.
This also needs a modification of the CONFIG options.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>
Tested-by: Wenyou Yang <wenyou.yang@atmel.com>