driver: at91_wdt: change the implementation to disable watchdog timer
From SAMA5D4, the watchdog timer is upgrated with a new feature,
which is describled as in the datasheet, "WDT_MR can be written
until a LOCKMR command is issued in WDT_CR".
That is to say, as long as the bootstrap and u-boot don't issue
a LOCKMR command, WDT_MR can be written more than once,
and the watchdog timer can be enable/disable without any constraints.

In current implementation, to disable the watchdog timer by only
setting the WDDIS bit(i.e. other bits are all set 0), this will
cause the watchdog timer can't be enabled in the future.
So it should be, read the WDT_MR register first, then modify the WDDIS
bit to disable the watchdog timer.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>