driver: pmc: BUG: clear the bit 13 of PMC_MCKR
From the support case, when booting kernel from Norflash
directly, the OHCI USB fails to probe during the kernel
boot up subsequently.

According to the information from the IC team, the bit 13
of PMC_MCKR register is used to control whether divide UPLL
clock or not. if it is 1, indicates divide the UPLL clock to half.

This bit is 1, after a processor reset. This bit has not
been overwritten before AT91Bootstrap when booting kernel
from Norflash.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>