Makefile: add the needed -mno-unaligned-access flag
Cross-compile with the gcc -mno-unaligned-access flag as unaligned accesses are
not allowed on a Strongly-Ordered memory.
As we don't configure the MMU (nor data caches, obviously), the core accesses the
memory in its configuration by default which is "Strongly-Ordered" (Cf. ARM
ARM, "B3.2.1 VMSA behavior when a stage 1 MMU is disabled".

It fixes the issues with SD card FAT compiled with newer gcc 5.2.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>