ddram: LP-DDR2: split sama5d3/sama5d4 from sama5d2 initialization
The newer sama5d2 DDR controller is different from the previous one that is
found on sama5d3 and sama5d4. Splitting the initialization sequence is the best
option to preserver the specificities and be aligned with datasheet step
description.

Just splitting for now: no actual modification nor functional change.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>