drivers: sdhc: fix card detection
The previous implementation was not so reliable: since MRL B, the internal
debounce time of the card detect pin has been increased, so the older
timeout value is now likely to be too small.
Also inconsistent combination of 3 bits in the PSR has already been
observed:
- 'card state stable': 1
- 'card inserted': 0 => a SD Card/eMMC has NOT been inserted yet.
- 'card detect pin level': 1 => the Card Detect pin is low level, hence a
                                SD Card/eMMC IS inserted.

This new implementation follows this sequence:
First we enable the 'card inserted' bit in the status register (NISTR).
Then we check the present state register (PSR) for the 'card inserted'
bit: a card has been inserted, this signal is debounced as opposed to the
'card detect pin level' signal. If the 'card inserted' bit is set in the
PSR, we stop here and claim a SD Card/eMMC is present.
Otherwise, we poll the NISTR for the 'card inserted' bit till either this
bit is set or we time out. Finally we read the NISTR one last time to
check the state of the 'card inserted' bit to claim whether a SD Card/eMMC
is present or not.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@atmel.com>