load_kernel: add encrypted kernel support
Linux kernel load is implemented as a 'load_function'. But load_kernel()
normally does not return to its caller. So the existing image decryption call
in main() is not applicable to kernel images. Add a kernel decryption step in
load_kernel() to allow boot of encrypted kernel images.

Tested on SAMA5D2 Xplained board.

Fixes #43.

Signed-off-by: Baruch Siach <baruch@tkos.co.il>