at91_aes: fix key and iv endianness
Secure boot tests on SAMA5D2 indicate that the KEYWRx and IVRx registers are
big-endian. This is documented in paragraph 57.4.1 (AES Register Endianness)
of the SAMA5D2 datasheet (rev. 11267E, 25-Jul-16). Switch endianness of the
values written to these registers. This allows expressing the .config key
values in natural hexadecimal writing, where the left most digit is the most
significant, and the right most is the least significant.

Signed-off-by: Baruch Siach <baruch@tkos.co.il>