Read Linux kernel boot arguments from file
This patch allow the at91bootstrap to read the boot arguments for the
kernel from a file in the same VFAT partition where the kernel is and
to send these boot arguments to the kernel.
This patch is for dual-bank application where the kernel is forced to
get the boot arguments from the bootstrap or where the user/app does
not have access/permission to change the kernel boot arguments.
I used gcc-linaro-4.9-2016.02-x86_64_arm-linux-gnueabihf to compile the
at91bootstrap together with this patch.
I test the binary on a custom board based on the SAMA5D27.

Signed-off-by: Yair Ben Avraham <yairba@tkos.co.il>