Makefile: fix gcc switch jumptable not linked with -fno-jump-tables
Previous commit added a switch statement. It looks on arm9 this instruction
links with libgcc jump table implementation in GCC. this means that in order
to have this working, it would be needed to link libgcc.a which might mean
increasing binary size.
Better solution is to have -fno-jump-tables to implement switch in place.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>