driver: sdhc: fix error on multiple block read above 8
When issuing a read command with more than 8 blocks we crash
because of the descriptor static array of size 8.
From the FS, we receive requests of 8 blocks(*512=4Kbytes) exactly
for each cluster. If the FS has a bigger cluster size, we receive 16 blocks
e.g. (16*512 =  8 Kbytes).

To avoid it, allow a multiple block read command with at most
16 blocks, which is line with most usecases. Increased descriptor table
length to 16. Also if someone tries to use a bigger cluster size, split
the command into separate 16 blocks commands to avoid any possible crash.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>