Fix recursive make call consistency in config-clean.
Every invokation of recursive make in Makefile uses $(MAKE)
except for the one in config-clean.  In BSD operating systems,
the default Make utility is POSIX Make, not GNU Make.  The
GNU version is in the ports tree and is named gmake.  By
making this replacement, at91bootstrap compiles cleanly in
OpenBSD.

Signed-off-by: Gregory P. Czerniak <greg@czerniak.info>