Fixes RTC precision loss in ACMESYSTEMS boards
https://www.acmesystems.it/ts_timelost

RTC time lost
Problem description
After setting the hardware RTC with the command:

    # hwclock -w
after just few minutes the RTC already lost seconds with the command:

    # hwclock -r
Solution
This problem is due to an error in the AT91Bootstrap configuration. We
missing to set this flag during the make menuconfig setup (see Compile
AT91Bootstrap).

The right configuration is this:

Slow Clock Configuration Options  --->
    [*] Use external 32KHZ oscillator as source of slow clock
    [ ]   Use external 32kHz clock as source of slow clock

Signed-off-by: Marcelo Roberto Jimenez <marcelo.jimenez@gmail.com>