board: sam9x60ek: remove nandflash_hw_init
Remove it because nand doesn't work. This was just a copy-paste
from sam9x5's nandflash_hw_init(). Even though the controller is
similar with the one from sam9x5, there are things that we must
take care of:
- EBI Chip Select Register is now in SFR,
- the pins should be set to default values,
- timings must be changed to match MT29F4G08BABWP nand flash
  requirements.
Since these changes summarize the entire nandflash_hw_init(),
remove it entirely to let room for a new implementation.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>