driver: sdhc: remove break condition on NISTR error after CR write
According to SDIO spec, after writing the CR (command register), the host
driver needs to wait for command completion.
It may happen that the NISTR register has an error initially, but once the
command completes, this error is replaced by command complete.
If at91bootstrap is compiled in THUMB mode, this scenario doesn't happen.
It is likely that in THUMB mode, instructions last long enough for this error
bit to not be read in time.
In ARM mode, it happens all the time. Adding a small delay before reading NISTR
fixes this.
However, the best practice would be to keep the timeout for the command
completion, and check the error bits at the end of the timeout period.
It can happen that the command completes in a timely fashion, in which case
the code will continue to work towards response read,etc, or the command is
never completed (error) in which case the normal behavior is not happening.
Either way, breaking the timeout waiting period immediately is not the good
way to implement the check.
Thus, removed the check for ERR bit from NISTR at every loop.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>