driver: twi: rework twi driver to support an array of controllers
This rework means rewriting the TWI driver to remove all ifdefs:
an array with the controllers' base addresses has been created.
A register function that will register a bus with the TWI driver
has been created.
This is called by the new register bus function, which needs
as a parameter the hardware initialization function.
The board hardware init is responsible to register the proper TWIs
with the driver.
This means that at the board init, the board, according to the SoC
and board capabilities, will register more or less TWI busses
with the driver.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>