board/sama5d27_som1_ek: Implement the VDDSDMMC over-consumption errata
For some of the SAMA5D2 DDR2 SiP, we must Implement the
VDDSDMMC power supply over-consumption errata.
We need to differentiate some of the SiP within the family and
apply a software workaround involving an analog SDHCI
calibration cell.
Do it in hw_init() function so that it's done at "cold boot" as well
as resume from BSR.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>