Config.in: allow the user to select eeprom twi if flexcoms available
In case the platform has only flexcoms, the board can autoselect the
twi bus (flexcom) to use to access twi.
However, if the autoconfig is disabled, the user cannot select anything
because the platform has no TWI.
This patch will allow to the user to select TWI0, TWI1, etc. as eeprom busses
even if they are in fact flexcoms (FLEXCOM0, FLEXCOM1, etc.)

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>