driver: sdhc: fix clock divider missing bits
Because of old typo, some bits of the clock divider were not being
written into the register. It happens for very high clock divider values.
Because of this, the frequency cannot go very low, as needed by some eMMC
devices (400kHz).
Fix the typo and shift the bits correctly into the register.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>