include: arch: pmc-v1: fix macros values
Commit 1965938c100b introduced pmc-v1.h header. In v2 of the patch series
that was send internally the content of pmc-v1.h was preserved from what
was defined in v1 of the patch series (internally) related to SAMA5D2.
But the pmc-v1.h introduced by commit 1965938c100b should have been
preserved the content of what was in include/arch/at91_pmc/pmc.h.

Fixes: 1965938c100b ("include: arch: at91_pmc: introduce a new header for 1st PMC version")
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>