driver: wdt2: introduce support for DWDT and WDTS
DWDT is the dual watchdog timer support in which one watchdog is
permanently not secure and the other one is programmable secure.
Introduce CONFIG_WDTS for SoCs that want to tackle this second watchdog.
Introduce CPU_HAS_DWDT for SoCs that support this feature
Add support for disabling this watchdog in the wdt2 driver.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>