Makefile: detect *nix shells
Detect if the current command interpreter is a Unix-like shell, or a
non-*nix interpreter such as Command Prompt/PowerShell.

When given a *nix shell, we'll further assume that all the necessary
tools are installed, and enable all the rules and features.
With a non-*nix interpreter we'll trim the recipes, skipping unsupported
commands and scripts.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>