Makefile: allow using specific GNU mkdir
Non-*nix interpreters such as Command Prompt may give precedence to
builtin commands over the GNU utilities sharing the same name, whatever
the priority of the latter is in the search PATH to executables.
This is typically the case with GNU mkdir, which may remain hidden under
the native MKDIR command.

On such build systems, one way of accessing the GNU utilities consists
in providing not only their name but the relative or absolute path to
them. Another and easier way consists in enclosing their name in double
quotation marks.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>