Makefile: support spaces in CROSS_COMPILE path
The CROSS_COMPILE environment variable may be assigned the path to the
cross toolchain, appended with the toolchain prefix e.g. 'arm-none-
eabi-' in the case of GNU Tools for Arm Embedded Processors.

On Windows systems, it is common for the cross toolchain to be
installed in a subdirectory of "<drive>:/Program Files/". Consequently,
paths to the tools there shall be enclosed in double quotation marks.

Ensure this is done for all cross toolchain invocations: gcc, as, ld,
objcopy.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>