driver: remove CONFIG_MEM_SIZE
We already have Kconfig options for CONFIG_RAM_32MB up to CONFIG_RAM_512MB
The CONFIG_MEM_SIZE is redundant, and moreover, when CONFIG_RAM_xxMB is
changed, the CONFIG_MEM_SIZE is left unchanged, which can lead to
misconfiguration.
With this change, once the user selects from menuconfig the appropriate
memory size, the MEM_SIZE is automatically reconfigured.
The code can still access MEM_SIZE, and it is now a compiler define
from Makefile.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>