driver: pmc: common: do not disable internal RC on AT91SAM9N12
Processor userreset hangs with internal RC disabled.

Atmel Errata:

51.1.3.1 Reset Hangs with 12 MHz RC Disabled

After power-up, the 12 MHz RC is enabled automatically. It is not
enabled after a reset, even if it is used during the boot
process. The user must not disable the 12 MHz RC to avoid any system
freezing on reset event.

Problem Fix/Workaround None.

Suggested-by: Ján Rusnák <jr@i-tel.sk>
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>