Support Off-Chip Memory Scrambling (OSMC) via config
SAMA5D3X boards have HW support for scrambling the
data bus to the external memory. This feature is
initialized with two 32-bit keys and an enable-bit.

This patch adds support to configure these keys
via Kconfig and extends the secure-driver with the
needed logic. The keys are handled like the other
'secure mode' keys to ensure confidentiality, thus
the feature is only available if secure mode is also
selected.

Signed-off-by: Kristof Havasi <kristof.havasi@gmail.com>