matrix: Pass arguments to matrix_read_slave_security()
Pass the base address of the Matrix together with the number of slaves
so that other chips can use this function too. As it were, these
arguments were hardcoded for the sama5d4 case. Also, the function
assumed that the system embeds two AHB bus matrixes, which is false
for the sama7g5: it embeds only one.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
[eugen.hristev@microchip.com: ported to at91bootstrap4]
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>