matrix: Get rid of local IDs and use matrix_base directly
matrix_configure_peri_security() and is_peripheral_secure() parsed
a static const struct peri_security array, retrieved a local ID
for the Matrix used, and based on that local ID got the base address
of the Matrix that was further used. The Matrix local ID is
superflous because it was used just for retrieving the Matrix's base
address. Drop it and use the base address directly. This also drops
the assumption that each system has 2 matrixes, which is not true
for sama7g5. AT91C_BASE_MATRIX32 and AT91C_BASE_MATRIX64 are SoC
specific, and on the sama7g5 one would have got the following errors:

/home/atudor/work/at91bootstrap/driver/matrix.c:549:10: error: ‘AT91C_BASE_MATRIX32’ undeclared (first use in this function); did you mean ‘AT91C_BASE_MATRIX’?
   base = AT91C_BASE_MATRIX32;
          ^~~~~~~~~~~~~~~~~~~
          AT91C_BASE_MATRIX
/home/atudor/work/at91bootstrap/driver/matrix.c:549:10: note: each undeclared identifier is reported only once for each function it appears in
/home/atudor/work/at91bootstrap/driver/matrix.c:551:10: error: ‘AT91C_BASE_MATRIX64’ undeclared (first use in this function); did you mean ‘AT91C_BASE_MATRIX32’?
   base = AT91C_BASE_MATRIX64;
          ^~~~~~~~~~~~~~~~~~~
          AT91C_BASE_MATRIX32

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>