pmecc: Rename GF Table offsets with meaningful names
No functional change intended. The Galois Field Tables offsets
were poorly named, not reachable through a grep when working
on pmecc. The Galois field tables are mapped in the ROM just after
the ROM code. Thus this is not a config option, strip "CONFIG" from
the name. Prepend PMECC to the name to indicate by whom these tables
are used. Indicate the size of the table in the name.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>