device: sama7g5: remove PA12 from SDMMC0 pinout
On sama7g5, PA12 can be used for SDMMC WP function, but on some boards it can
serve a different purpose. Remove the setting of this pin to SDMMC function.
Even if it is connected to the WP of the card, it is useless in bootstrap
context as we do not perform writes on the card.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>