driver: backup: add infrastructure for ddr phy recalibration
When entering to backup and self-refresh mode via Linux, Linux saves
the calibration information and first 8 words of DRAM to securam. In
the process of DDR PHY reconfiguration (while exiting backup and
self-refresh mode) the code is using these information for
reconfiguring the DDR PHY. The first 8 words of DRAM are necessary
because these might get corrupted in the process of PHY recalibration
and thus needs to be restored to DRAM afterwards.

Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>