Makefile: add extra cc args as external parameter
Add the EXTRA_CC_ARGS which can be used to pass extra parameters to the
toolchain.
By default this is empty.

This is done to address issues introduced by
bc526cc5 ("Makefile: support spaces in CROSS_COMPILE path")
which added quotes to the $(CC) $(AS) and $(LD)

Yocto project sets the $(CC) to 'gcc --sysroot=/...'
and with the quotes in place, the shell will try to invoke the command
'gcc --sysroot...' considering the space and sysroot as part of the command
instead of being a parameter to the 'gcc'. This will cause build failures
with Yocto Project.

With the EXTRA_CC_ARGS, Yocto can pass the sysroot in this parameter and
keep the $(CC) clean for the quotes to work correctly.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Reviewed-by: Aubin Constans <aubin.constans@microchip.com>