device: sama7g5: add hw preinit for lowlevel_clock_init
SAMA7G5 requires a preinitialization for the clock scheme.
Out of Rom Code, the PLLs are already initialized.
We have to reconfigure them earlier in the boot process, before
changing the source of main clock to the external crystal.
If we change to the external crystal before reconfiguring the PLLs,
there is a slight chance that the PLLs are blocked.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>