device: sama5d3: remove board_override_cmd_line()
The Android-specific code in this function is not needed anymore.
Then, this version of the function can be
removed, and we can instead rely on the default, weak implementation
defined in driver/load_kernel.c.

Incidentally, this fixes a bug, as this board_override_cmd_line() would
return NULL unless CONFIG_LOAD_ANDROID was set.

Signed-off-by: Edgar Bonet <bonet@grenoble.cnrs.fr>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>