driver: Config.in.nvm: introduce CONFIG_SDHC_8BIT_SUPPORT
SAMA5D2 device has 8 bit SDHC0. On some boards though, on SDHC0 is connected
to an SD-Card , which only uses 4 pins. In this case the other 4 pins are
used to another device, for example the wlsom1_ek which uses them for the LEDs.
Introduced CONFIG_SDHC_8BIT_SUPPORT, which is by default yes, and allows to
deselect the 8 bit support, and have AT91Bootstrap initialize just 4 pins.
This will allow pins to be used by LEDs for example.

Reported-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>