Add scripts and project definition for MPLAB X IDE
Add a NetBeans Project that the Microchip MPLAB X IDE will recognize as
a valid MPLAB X Project.
The MPLAB X -defined type of this Project is "User Makefile".

Add MPLAB X -specific Python scripts, which implement, notably, a
"Kconfig" panel in the Project Properties. This GUI, as a frontend to
Kconfiglib, offers features equivalent to the 'defconfig' and 'xconfig'
Makefile rules.
Note that MPLAB X uses Kconfiglib and not the Kconfig tools.
Consequently, it is able to configure AT91Bootstrap without resorting
to the command line.

All Python scripts here are meant to be interpreted by Jython 2.7.

Signed-off-by: LI BIN <bin.li@microchip.com>