pmc: clk-common: introduce CONFIG_MAINOSC_MEASUREMENT
CONFIG_MAINOSC_MEASUREMENT allows the lowlevel_clock_init code source to
perform a measurement of the main crystal oscillator.
This was previously done on sama5d2 and sama7g5, but on sama7g5 the measurement
result was being ignored.
Implemented the measurement and compared with datasheet >12 Mhz <48 Mhz .
Because the measurement is done with MD_SLCK domain, it's pretty inaccurate.
Relaxed the conditions with >10 Mhz and <50 Mhz

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>