device: sama7g5: split the CPU cruise speed configuration from hw_init
Split the hw_init function into two parts. One old part will configure
everything except for the CPU cruise speed. It will continue to set the CPU
speed at the safe 600 Mhz speed.
The new function hw_postinit will handle the CPU Clock selection for the cruise
speed.
The purpose of this split is that this new function must be called after the
PMIC voltage selection function which configures the recommended VDDCPU value
from the datasheet.
The device/sama7g5 file should be agnostic of the PMIC used on the board.
Thus the PMIC needs to be configured at the bootloader level, not at the
device level.

Signed-off-by: Mihai Sain <mihai.sain@microchip.com>