configs: sama5d2_xplained: add RGB LED support and set LED GREEN on
The LEDs on this board are connected to PMIC OUT6. The schematic of the
board assumes OUT6 at 3.3V.
In consequence, correct the output of OUT6 to be 3.3V.
Add RGB LED support with LED RED and BLUE set off.
Enable LED GREEN at boot time.
As on the other boards, green led indicates at91bootstrap started.

Signed-off-by: Mihai Sain <mihai.sain@microchip.com>