driver: sdcard: do not load device tree or cmdline if pointers are NULL
If Linux is selected as the next software to load, the load_sdcard function
will always try to load a device tree and a cmdline file if enabled. With OP-TEE
support, the code reuse this function to load the OP-TEE image. However, we do
not have any device tree or cmdline args to load. Add checks for the destination
address to be non NULL. In the same time, fix a wrong check on sdcard_read_cmd
return value.

Signed-off-by: Clément Léger <clement.leger@bootlin.com>