device: sama7g5: Add SMC NAND support
Add SMC NAND support. The sama7g5's SMC IP is the same as sama5d2's with
a slightly change: it provides a synchronous clock output (SMC clock)
that is dedicated to FPGA usage. We don't address the FPGA use case as of
now, so don't touch the SMC clock. Tested with the 29F4G08ABAEA NAND
flash and could fetch a u-boot binary. HW PMECC was exercised and could
correct 8 bitflips.

Configure the Matrix for SMC EBI CS3 and NFC RAM. EBICS3 can map up to
128 MBytes. Configure the Region Top Address at maximum value (128 MB),
so that SMC can use the full 128 MBytes on the system buys. Split the
region at Region Top Address offset and configure the Low Area as
Not Secured, so that u-boot and kernel can access the full 128 MBytes on
EBICS3. Similar configuration is done for NFC RAM, but with a different
size: 16 KBytes. NFC SRAM should be mapped on 9 Kbytes, but there's no
such granularity to limit the access above 9 KBytes, so use the closest
minimum granularity (16 KBytes).

Configure the NAND pins, enable the HSMC peripheral clock, configure the
SMC timmings, and define the PMECC offsets.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>