scripts: addpmecchead.py: Allow binaries without PMECC header
In case one uses CONFIG_PMECC_AUTO_DETECT to discover the PMECC
parameters, addpmecchead.py is in the imposibility to get the ONFI
discovered params, as those are read at runtime. To address this,
addpmecchead.py uses default PMECC parameters for each SoC.
This is not very flexible, as one may choose to test different NAND
flashes with different PMECC caps with the same SoC, thus we shouldn't
generate a PMECC header per SoC, but rather per flash or board if really
needed. Thus I choose to not define default PMECC params for sama7g5, but
also because the sama7g5ek does not populate a NAND flash. As it was,
if one used CONFIG_PMECC_AUTO_DETECT and the SoC did not define default
PMECC parameters, the build failed. Be more permisive and allow binaries
without PMECC header. In such cases, $(BINDIR)/pmecc.tmp and
$(AT91BOOTSTRAP).pmecc will not be generated.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>