mci_media: fix fallback to 1 bit bus mode
By default assume that if 8 bit and 4 bit bus modes do not work, fall back
to 1 bit bus mode.
The select bus width function should be able to be called with this parameter.
Also, when setting DDR, avoid setting DDR if the buswidth is 1, because
for DDR, only 8 bits and 4 bits are supported.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>