configs: sama7g5: change core frequency to 800 MHz by default
The core frequency of 800 MHz is recommended for best lifetime and best
power consumption and operation of this device.
The frequency of 1000 MHz can be used when more computing power is required.
If at91bootstrap is booting linux, the DVFS features of Linux will take over
and increase/decrease the frequency as per OS needs.
Note that the 1000 MHz frequency requires certain operating voltages to the
MPU core, and this is board specific feature.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>