Makefile: fix stack space message computation
The stack space is the total SRAM area minus the size of the bootstrap binary.
The total SRAM area is double the maximum size for the bootstrap binary.
When the Romcode copies the next stage into SRAM, it can only copy a maximum
size equal with half the SRAM. Hence the total SRAM is double the maximum
size for the bootstrap binary.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>