configs: sama7g5: remove MCP16502 initialization
With the patch
fd21ad62 ("configs: sama7g5: change core frequency to 800 MHz by default")
the core of the sama7g5 no longer requires an increased voltage for the
VDDCPU.
Thus, it's no longer required to initialize the MCP16502 on the board
and change it's default voltage outputs. The defaults are already configured
for 800 MHz operations.

Fixes: fd21ad62 ("configs: sama7g5: change core frequency to 800 MHz by default")
Reported-by: Mihai Sain <mihai.sain@microchip.com>
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>